pyproject.toml
[project]
name = "lontod"
version = "0.1.0"
description = "A daemon to serve ontologies in a linked open data approach"
authors = [
    {name = "Tom Wiesing",email = "tkw01536@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "rdflib (>=7.1.3,<8.0.0)",
    "starlette (>=0.46.1,<0.47.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "html-sanitizer (>=2.5.0,<3.0.0)",
    "python-mimeparse (>=2.0.0,<3.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "python-debouncer (>=0.1.5,<0.2.0)",
    "pip-licenses-lib (>=0.5.0,<0.6.0)",
    "dominate (>=2.9.1,<3.0.0)",
    "markdown (>=3.8,<4.0)",
    "types-markdown (>=3.8.0.20250415,<4.0.0.0)",
]

[project.scripts]
lontod_index = 'lontod.cli.index:main'
lontod_server = 'lontod.cli.server:main'

[tool.poe.tasks]

format = [
    "_ruff_format",
    "_ruff_fix"
]
_ruff_format = "ruff format"
_ruff_fix = "ruff check --fix --unsafe-fixes"

lint = [
    "_lint_ruff",
    "_lint_pylint",
    "_lint_mypy"
]
_lint_ruff = "ruff check"
_lint_pylint = "pylint src/ tests/"
_lint_mypy = "mypy src/ tests/"

test = [
    "_test_pytest"
]
_test_pytest = "pytest"

ci = [
    "lint",
    "test"
]

[tool.pylint]
disable = [
    "C0301", "E1121", "R0801", "R0902", "R0903", "R0911", "R0912", "R0913", "R0914", "R0915", "R0917", "R1702", "W0511", "W0718",
    "W0611", # ignore unused imports -these should be fixed by 
]

[tool.ruff]
src = ["src", "tests"]

[tool.ruff.lint]
extend-select = [
    "YTT", 
    "ANN",
    "ASYNC",
    "S",
    "BLE",
    # "FBT",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "EM",
    "EXE",
    "FA",
    "INT",
    "ISC",
    "ICN",
    "LOG",
    "G",
    "INP",
    "PIE",
    "T20",
    "PYI",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SIM",
    "SLOT",
    "TID",
    "TD",
    "TC",
    "ARG",
    "PTH",
    "FLY",
    "I",
    "C90",
    "NPY",
    "PD",
    "N",
    "PERF",
    "E",
    "W",
    # "DOC",
    "D",
    "F",
    "PGH",
    "PL",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "UP",
    "FURB",
    "RUF",
    "TRY",
]
ignore = [
    "ANN401",
    "BLE001",
    "COM812",
    "E501",
    "TD002",
    "TD003",
    "D203",
    "D213",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "S101",
    "INP001",
]

# TODO: remove these when the rework is finished
"src/lontod/html/*" = [
    "C901",
    "PLR0913",
    "PLR0912",
    "D205",
    "PLR0915",
    "PLR0911",
]


[tool.poetry]
packages = [{include = "lontod", from = "src"}]

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
pylint = "^3.3.5"
pytest = "^8.3.5"
poethepoet = "^0.33.1"
ruff = "^0.12.0"

[tool.mypy]
mypy_path = "typings/"
strict = true

[[tool.mypy.overrides]]
module = ["html_sanitizer.*", "dominate.*"]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
